[Refactor] modularization
